RECONORNE/requirements.txt
requests
whois
dnspython
beautifulsoup4
exifread
colorama
